Update patch set 1
On 2008/11/02 02:04:40, "Evan JIANG" <firstfan wrote:
> And to his question, I've check the saving function code, it's ok.
This
> patch will make some of the contacts which are stored before wrong.
But
> not big problem, only the organization label is changed. Users only
need
> to set that again.
> It's better to make this change than much later. The type definition
and
> the type label are different. I cannot think out a solution that could
> solve this problem perfectly.
Oh ok. I got confused by a fact that when editing it presented
non-existing 4th choice (home). Now I see that is simply pulls the list
of the strings from resources, so fixing the resource shouldn't affect
saving code.

As for the label change I understand the motive and also think it would
be better done now than later. However it does introduce a slight
problem to be - I'm working on a contact import/export application,
which makes the data produced by it incorrectly either before or after
this bug fixed. Well I guess I'll have to cope with it somehow :)

Patch-set: 1