Update patch set 1
No score.

I will second swetland -- "Android on Freerunner support" certainly does
not accurately describe the bulk of what is here.  Most of this looks
like very separable individual changes that are really not specific to
Freerunner and should be submitted as specific patches with descriptions
of what they are doing.

Patch-set: 1