Update patch set 1
No score.

It is not necessarily a good idea to center the boot animation. Another
option is to scale it up or down (with proper filtering). There are 2
reasons fore this:

1) at higher resolutiona the screen's dpi may be high and the animation
may appear too small.

2) if the background of the image is not black, it will not match the
background and look ugly.


This change doesn't break anything in the case the screen is too big or
too small for the assets and improves things in some cases, so I'm not
scoring negatively; however it doesn't fix the problem fully.

Patch-set: 1