Update patch set 1
Mathias,

I agree with you. It doesn't fix problem fully. But I think it's worth
merging it for non-HVGA panel users who wants to see robot glowing
animation. It looks good on VGA or W-VGA panel. In the high resolution
screen case, it should be made a new animation which suited for the
screen size.

-Jyunji

On 2008/11/03 06:39:51, "Mathias Agopian" <pixelflinger wrote:
> No score.
> 
> It is not necessarily a good idea to center the boot animation.
Another
> option is to scale it up or down (with proper filtering). There are 2
> reasons fore this:
> 
> 1) at higher resolutiona the screen's dpi may be high and the
animation
> may appear too small.
> 
> 2) if the background of the image is not black, it will not match the
> background and look ugly.
> 
> 
> This change doesn't break anything in the case the screen is too big
or
> too small for the assets and improves things in some cases, so I'm not
> scoring negatively; however it doesn't fix the problem fully.
> 
> 
> 

Patch-set: 1