Update patch set 1
I would prefer that you didn't submit this.

This change touches the files

core/java/android/provider/Contacts.java and 
core/res/res/values/strings.xml

without saying so in the commit.
These two files are not part of the changes in the parents.
Thus this is a merge commit that introduces changes that were not in its
parents, an "evil merge".

Since such changes are hard to track and review in git alone and even
harder in Gerrit i would propose you put them in a separate commit,
along with an appropriate commit message.

Patch-set: 1