Switch long date format to something more human friendly:
      December 09, 2008 --> December 9, 2008
      09 December, 2008 --> 9 December, 2008

This human visible date is prominently displayed on the status bar
when you roll it down. Right now, it says "December 12, 2008" which is
soothing and relaxing. However, a few days ago it said "December 09,
2008" in which  the leading zero caused myself, and presumably others,
a small amount of mental discord.  Put another way, this change will
the end user experience slightly better 30% of the time.