Update patch set 1
Looks good to me, but someone else must approve.

Issue #1324 involves a disparity between the in-call voice audio and
in-call generated audio.  In particular, when the in-call volume is at
the minimum, the voice audio is audible, but call-waiting notifications
are not.  This means that users who find the in-call volume to be most
comfortable at its lowest level do not hear call-waiting tones.

The disparity arises because the voice call audio is routed in hardware,
whereas generated audio such as call-waiting notifications are handled
in software in AudioFlinger.  In all of the Android volume scales, 0.0
(or 0) means inaudible and 1.0 or (100) means full.  However the
hardware voice call audio level is never inaudible, and 0 means instead
the lowest audible level.  Unfortunately, the single audio stream
STREAM_VOICE_CALL is used for both the hardware volume and in-call
generated audio volume.  This disparity is already being dealt with
elsewhere in Android; consider this comment from
frameworks/base/core/java/android/view/VolumePanel.java:

    * For in-call voice call volume, there is no inaudible volume
    * level, so never show the mute icon

The attached change causes in-call audio never to be inaudible as well. 
In AudioFlinger, it scales the selected volume level and shifts it by an
offset so that there are still 6 levels of user selected volume, but the
lowest level is now audible.  It essentially translates the Android
volume scale to match the in-call hardware volume scale.

Patch-set: 1