Update patch set 2
No score.

I doubt it is device specific.  Elsewhere in the framework, the
assumption that underlies the problem is expressed:

frameworks/base/core/java/android/view/VolumePanel.java:

    * For in-call voice call volume, there is no inaudible volume
    * level, so never show the mute icon

Based on that, it seems that there is already an assumption in the
framework that, unlike other volume channels, in-call volume is never
muted.  This patch brings in-call generated audio in line with in-call
voice audio.

Patch-set: 2