Update patch set 2
Patch Set 2: No score

I think you are misunderstanding the purpose of in-call audio stream type. There should never be any generated streams that use this type. In fact, there is no way on any current Android devices to stream audio generated on the apps processor to the uplink audio circuit.

This is just an artifact of some function overloading in the framework. We plan to re-architect this part of the code soon to deal with some other issues that are coming up related to SCO, A2DP, and other audio routing issues.

Patch-set: 2
Reviewer: Gerrit User 1000387 <1000387@85c56323-6fa9-3386-8a01-6480fb634889>