Update patch set 2
Patch Set 2: No score

I am sorry that I haven't been able to communicate this effectively, but I do believe I understand how audio in this case works from top to bottom.  Moreover, I have tested this patch on my phone and it works as intended.

The type of audio I'm concerned with is not generated audio sent to the uplink circuit (which I take to mean the audio sent from the phone to the cell tower), but rather audio generated on the apps processor and sent to the local speaker on the phone (whichever output that may be).  The call waiting tones are my specific concern, but call completion tones and busy signals are another example.

The call waiting tones start in:
  packages/apps/Phone/src/com/android/phone/CallNotifier.java:
   toneGenerator = new ToneGenerator(AudioManager.STREAM_VOICE_CALL, toneVolume);

Where you can see that the stream type is STREAM_VOICE_CALL, which is the Java equivalent of the C++ AudioTrack::VOICE_CALL.  Also, to lengthen the citation of my earlier example, here is the special case handling of VOICE_CALL in the volume control panel:

  frameworks/base/core/java/android/view/VolumePanel.java:
            case AudioManager.STREAM_VOICE_CALL: {
                message = INCALL_VOLUME_TEXT;
                /*
                 * For in-call voice call volume, there is no inaudible volume
                 * level, so never show the mute icon
                 */
                setOtherIcon(index == 0 ? 1 : index);
                break;
            }

Indicating both that this is used, and that the framework currently deals with the VOICE_CALL stream specially.

The bug is that when the VOICE_CALL stream volume is set at 0, all generated in-call sounds are muted.  But that stream volume is overloaded and also controls the (hardware routed) in-call audio volume (how loud the phone call sounds to the user).  That audio is never muted, even at 0, so the sounds controlled by this one audio stream level are disjoint.  The volume panel handles this by indicating that the channel is not muted, even when at 0, which is the behavior of the hardware audio.  My patch makes the behavior of the generated audio match that by also never being muted.

I wrote above that I have tested this patch.  Here is a test procedure:

1) Make a phone call to tie up the line.

2) Set the in-call volume to something >0, so there is a yellow bar in the volume panel.

3) Call in to the phone; you should hear call waiting tones on the phone.  Don't answer.

4) Set the in-call volume to 0, it's lowest.

5) Call in again.  Without the patch, you will not hear call waiting tones.  With this patch, you will.

Patch-set: 2
Reviewer: Gerrit User 1000341 <1000341@85c56323-6fa9-3386-8a01-6480fb634889>