Update patch set 2
My apologies. I had no idea we were using the IN-CALL volume for app processor streams. That is a bug.

I'll accept the patch if you add a FIXME indicating that this is a temporary patch. All the audio stream management is going to get a rework in a future release and this problem needs to be resolved in a cleaner way.

Patch-set: 2