Update patch set 1
Patch Set 1: No score

This is a OTA Provisioning Client for android. It enables the use of OMA 1.1 Client Provisioning (via WAP PUSH) to set up APNs on the device.

It depends on Change 8272 to implement the content type security handling.

Sorry about the above description, I cannot find out how to retract this commit to change it in this version of Gerrit.

Patch-set: 1