Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

The current implementation of AudioRecordThread loop has a
bug in that if getNextBuffer() returns a size less than the hardware input buffer size we will trigger a read error and data will be lost.

However we would prefer to keep the implementation of RecordTrack::getNextBuffer() generic as it is also used for Track::getNextBuffer(). We want to keep it suitable for the future when we are not limited to read blocks of fixed size.

We would rather fix the code in AudioRecordThread loop by testing the size of the returned buffer and not read from hardware input if less than imposed buffer size with a comment indicating why we do so. This is slightly less efficient but makes the reason of this limitation
more clear.

Our internal version of Cupcake has this patch and we will push it out when we merge the rest of our changes.

Thanks for your help in finding this bug!

Patch-set: 2