CameraService: Reuse the overlay when preview is restarted after image capture
In CameraService::Client::startPreview(), an instance of Overlay is created
and it is passed to the CameraHal. Before sending a command to capture an
image, the preview is not stopped by the application. Therefore this overlay
object is not freed. After image capture, the application once again calls
startPreview(). At this point, the CameraService tries to instantiate
another instance of Overlay. The previous instance of Overlay cannot be
freed because the destructor is private.

Therefore, I modified CameraService to reuse the same instance of Overlay.

Signed-off-by: Anu Sundararajan <sanuradha@ti.com>