Update patch set 3
Update to Description:
The problem was spotted with broadcasting an ACTION_NEW_OUTGOING_CALL intent which currently does not
require any permissions. It is suggested to require the CALL_PHONE permission for the
broadcast receiver in Phone application (and maybe also add the uses-permission CALL_PHONE to AndroidManifest of Phone application).

So this issue hid another bug. And with this patch the change of Phone application would become necessary to avoid unauthorized calls.

Patch-set: 3