Update patch set 1
Patch Set 1: No score; No score

I haven't had a chance to look at this in detail, but one thing I think needs to be improved is all the code spread around to tell with the new setting -- we shouldn't be duplicating that code, there should be one API that returns the correct time zone.

Shouldn't this just be setting the current time zone to whatever the default is, so we don't need special cases in the alarm manager and elsewhere to deal with this new option?

Patch-set: 1
Reviewer: Gerrit User 1000559 <1000559@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=0