Update patch set 1
Thanks for the response, Dianne.

Can you be more specific on the duplication?  There is a single API and thats the Setting.System.getString()  
Timezone is initialized in AlarmManagerService, and used in one other place that was relying on the old auto_time flag.

Rest of the changes are customary for the new Settings.  Is there a better way? (I'm using the implementation of auto_time flag as reference)

Patch-set: 1