Update patch set 1
Dianne,
The ServiceStateTracker.java was included in the change because I noticed the getAutoTime() being used in the same code-block.  Now I realize that the added code is not necessary when autotime is not set at all.  I should remove the ServiceStateTracker.java change.

Besides, my change to AlarmServiceManager is outdated now.  I have to stop calling the setTimeZone() from the constructor, because it tries to broadcast (and it fails as the construction happens during BOOT_COMPLETED time)  So, I would change the setTimeZone() call to equivalent to  setTimeZoneNoBroadcast. (and make setTimeZone() call the setTimeZoneNoBroadcast + then do the Broadcast)

Would the above suffice?  Can you please update, so I can abandon this change and redo this again.

Also, Am I supposed to work on the donut branch to make this change reach the next release?  

thanks,
Jey

Patch-set: 1