Add option to set ListPreference summaries from entryValues
Currently when ListPreferences are used in a PreferenceActivity, the summary
values are set to the same as the current index in mEntryValue. This patch
add an option "summaryFromEntries" to allow the summary to be set from the
corresponding index in mEntries to aid in localization.

For example a preference may be named "color" with the following attributes
in the locale "de" (German):

mEntryValues = { "red", "green", "blue" }
mEntries = { "rot", "grün", "blau" }

Change-Id: Iad9e96a49f1266d20f537f5fc77ca1a7bd284a84