Fix intermittent deadlocks during boot, observed on a hardware platform.
Since the variable read is atomic, the only effect of the
synchronized() call would be to block the calling thread until the
holder of the lock releases it, and it is possible that some code path
might depend on this side-effect.  But testing using the cupcake
branch running on hardware and emulator (by many users) showed no
problems with the change.  Review by an expert in the code may
nevertheless be in order.