Update patch set 1
Patch Set 1: No score; I would prefer that you didn't submit this

The performance impact of this change is too big. I don't understand why it is needed. Maybe it is a particularity of the U8500, in this case this change should be made conditional.

Patch-set: 1
Label: Code-Review=-1
Label: Verified=0