Update patch set 3
Patch Set 3: No score

Shrinks generic-eng system.img from 51,395,520 to 49,953,024 for a savings of 1,442,496 bytes.

For a generic-eng-{en_US,es_US} the size goes from 46371072 to 46,178,880 for a savings of 192,192 bytes.

XML will take some work to make sure we're not wastefully converting back and forth.

Patch-set: 3
Label: Code-Review=0