Update patch set 2
Patch Set 2: Verified; I would prefer that you didn't submit this

Patch Set 2 is a clean cherry-pick of patch set 1 on top of the latest master (to avoid a merge path conflict).

I never feel comfortable seeing a fix for a deadlock that removes a lock without at least a thorough analysis of the situation that causes the deadlock.

Patch-set: 2