Update patch set 2
Patch Set 2: No score; Do not submit

Agreed.  This needs to include the stack crawls of the threads involved in the deadlock and show why this is the correct fix.  I think it is very unlikely to be a correct fix, because a basic tenant of the threading model is that the window manager is free to acquire its main lock while handling any incoming call.

Patch-set: 2
Label: Code-Review=-2
Label: Verified=0