Update patch set 5
Patch Set 5: I would prefer that you didn't submit this

Hi Kenny,

I have a couple suggestions:

Another more flexible avenue is to allow the summary to be a template that includes a replacement variable for the current value's entry.  For example, using that approach, you could set the summary to "%1$s" and it would always be the current value's entry (equivalent to your existing approach does).  If you wanted to get more fancy, you could have the summary be "The color is %$1s".

Also, it would be useful to add this functionality to EditTextPreference as well.

What do you all think?

Patch-set: 5