Update patch set 4
Patch Set 4: No score

Found a few existing bugs with ResXMLTree handling and fixed those so now it works with UTF-8.

The only thing missing now is changing ResStringPool::indexOfString to convert the target key to UTF-8 before searching through a UTF-8 string pool instead of converting each of the values in the pool to UTF-16 for comparison.

Patch-set: 4
Label: Code-Review=0