Update patch set 5
Patch Set 5: Looks good to me, but someone else must approve

Pushed the previous version with a new commit message and a TODO in indexOfString noting that there could possibly be an efficiency improvement. I believe this is my final answer.

As a note about the TODO mentioned: I changed about 200 lines trying to get a good, clean improvement from indexOfString, but I don't think the complexity of the change is worth it. It doesn't appear to be used often at all.

As it stands the conversions will be cached as it's doing a binary search which will be a tiny impact even for the biggest resource pools. For instance, Settings will have about 11 iterations to find the key. Each search will share some of the midpoint comparisons, so they'll be cached after the first couple of searches anyway.

Patch-set: 5
Reviewer: Gerrit User 1000748 <1000748@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1