Add ViewInvalidationObserver to View class.
This makes it possible for a program to be notified when a view is invalidated
and needs to be redrawn. Can be used for example when drawing views to
off-screen buffers and using them on opengl surfaces. Without this feature
you have to rerender view off-screen every frame to be sure you dont miss any
updates.

(Have example of list on 3d surface up and running using this patch).

Update: For effieciency. removed iterator + only one if statement evaluated in
invalidate() if no listener present.