Solution for issue of inflating big compressed files.
There is a size limitation on buffer for inflating a compressed file.
The the file is bigger than 1MB, ZipUtils does not allocate the buffer
for inflating.  This changeset provide the ability to inflate a file
into a file descriibed by a FILE pointer.  And, user of ZipUtils no
more limited by size limitation of buffer.

Change-Id: Ie13671e6cbeb5d66ec340f6d07aa72d8dd3c99f5