Fix the issue of loading big resources.arsc.
When aapt try to load resources from resources.arsc in android.jar, it
would be fault for resources.arsc that is too big to squashed into a
1MB (or 2MB) buffer when inflates it.  The solution is inflating the
file into a temporary file.  The content of resources.arsc is loaded
in several times to avoid to allocate a large block of memory.
Everytime, it loads one or more complete chunks and squash chunks into
a buffer (2MB).  These chunks are add into ResTable in several times
to avoid allocation of big memory blocks.

Change-Id: Ieb50818ee7b8bbeb18fe4f142ba642d63af67f9d