Update patch set 2
Patch Set 2: No score

That's essentially what the original code is doing (using the alphanumeric check as a fallback).  In most cases gsmRoaming will be correctly reported by the baseband - the baseband will determine this based on MCC/MNC compared against a list stored on SIM (and possibly an additional database maintained by the OEM).  The alpha check is really there to cover the case where the baseband doesn't know about a newly acquired network.

I'd be really surprised if gsmRoaming were true AND the numerics matched, and would consider that a bug in the baseband.

Patch-set: 2