Update patch set 2
Okay, that makes sense now. I did not think of the baseband already checking the numeric ids. The reason for changing this piece of code was that my SIM does not seem to provide any alphanumeric id. So in short: my logic was wrong, thanks for pointing this out.

Patch-set: 2