Update patch set 2
I'm not sure there's a whole lot we can do for that case.  The code does not set roaming == true just because the alpha is missing on the SIM - the baseband has to indicate that the device is roaming as well.  If alpha is not available to check, I don't know of any other fields that would be appropriate.

Patch-set: 2