telephony: Fix for MMI code for call forwarding
Fix for issue - The UI doesn't register for call forwarding when
*SC*SI# is dialed where SI is the number to which call should
be forwarded.

3GPP specification 22.030 sec. 6.5.2

The UE shall determine from the context whether,
an entry of a single *, activation  or registration
was intended.

For example, a call forwarding request with a single *
would be interpreted as registration if containing a
forwarded-to number, or an activation if not.

Change-Id: I0531a134802f807a9b3023e69393bd5726ed1129