Update patch set 1
Patch Set 1: No score

Sure, well, the issue was that when you can digits in RTL text, the digits would show in RTL, instead of in LTR order as they should.

It seems that the original java set the direction of digits to the direction of the text before it, and it should always be LTR, so i changed that.

Patch-set: 1
Reviewer: Gerrit User 1002999 <1002999@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0