Update patch set 1
Patch Set 1: No score; Do not submit

What Dianne said. 

If you wanted to change the notification manager and the HAL API to set the battery light to the battery level with 0..255 in each of the RGB channels, and put all of the logic for the rest of the blinking and color into the HAL, I'd be okay with that.

Patch-set: 1
Label: Code-Review=-2
Label: Verified=0