New SurfaceFlinger API to allow grabbing screen shots
This patch adds a new interface through which clients can obtain
images of the screen from the SurfaceFlinger. This is needed because
reading directly from the frame buffer is unreliable on some
platforms.

The new entry point is

status_t SurfaceComposerClient::grabScreen(DisplayID dpy, int fd);

To use this API, the client should create an ashmem region of the
correct size (which it can calculate from the data returned by
getDisplayInfo) and pass the file descriptor to this function. The
call will fail unless the client has READ_FRAME_BUFFER permission.  It
will also fail if a surface with the SECURE flag set is visible.

On return the ashmem region will contain a copy of the current screen
image (in OpenGL row order).

Future patches will build on this API to provide more features such as
incremental screen capturing (grabbing only changed regions, rather
than the whole screen) and the ability to grab only parts of the
screen, rather than the whole image.