Update patch set 1
Patch Set 1: No score

I don't understand why this is necessary.
The only strong reference on the AudioFlinger TrackHandle (IAudioTrack) is held by the client process. So when it dies the TrackHandle should be destroyed and the track stopped.
As far as I remember, this was working. I will check again, maybe there is something special when using STATIC mode.

Patch-set: 1