Update patch set 1
Patch Set 1:

I just made the test with a streaming AudioTrack and the TrackHandle destructor is called when the client process is killed. It is destroyed immediately, not waiting for track underrun. Moreover, the binderDied() callback is called in AudioFlinger.
There must be something specific with the static mode.

Patch-set: 1