Update patch set 1
Patch Set 1:

The problem is a bug in AudioFlinger that makes that even if the TrackHandle is destroyed, if a track is active and has frames ready for mixing, it will not be destroyed.
If the track is in static mode and the sound is looped, this track will play for ever.
There si a simple fix in prepareTracks_l().

Patch-set: 1