Update patch set 1
Patch Set 1: No score; I would prefer that you didn't submit this

It is a reasonable change, especially if we add in settings to use Active power mode option. However, "driver getpower" is implemented only for wl1251/wl1271 drivers and will get unpredictable results for bcm4329. It can be pushed when we add this command support to the driver.

Patch-set: 1
Label: Code-Review=-1
Label: Verified=0