Update patch set 1
Patch Set 1:

This fails to merge, and one of the reasons is that a file disappeared since this change was uploaded, possibly in order to support CDMA devices.

If this change is still relevant, please resolve the conflicts and upload a replacement.

Otherwise it's probably a good candidate to be marked as abandoned.

Patch-set: 1