Update patch set 2
Patch Set 2: No score

Hi,

Here is further motivation for this change from the team behind it.

When a key is pressed for a long time, the Key Event will be repeated until the key has been released by the user. This key repetition can be done in either the kernel or in the WindowManagerService. 

The KeyEvent contains a method named getRepeatCount which is used to retrieve the repeat count of the event. This is the number of times the key has repeated.

The problem is that this function only works if the repetition was done in the WindowManagerService. The repeat count will always be 0 if the KeyEvent was repeated in the kernel. This fix solves this problem.

The problem can be reproduced with a bluetooth-accessory with fast forward and rewind. The repeat count in the KeyEvent will always be 0 when these buttons are pressed for long time.

The same motivation applies to the other part of this change in https://review.source.android.com//#change,13636.

Patch-set: 2
Reviewer: Gerrit User 1002951 <1002951@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0