Fix updating and deleting FDN entries with an empty alpha identifier.
- Fix AdnRecord.buildAdnString() to generate the correct record when alpha
identifier is empty. This allows the user to update an FDN entry to remove
the alpha identifier. Previously the entire entry would be deleted because
an empty record was generated here when the alpha identifier was empty,
rather than a record containing the phone number with an empty alpha tag.

- Fix IccProvider.delete() to construct an AdnRecord with "" for the alpha
identifier instead of null. This prevents a NullPointerException in
AdnRecord.isEqual() when deleting an FDN record with no alpha identifier.
See the comments on https://review.source.android.com//#change,13492
The original version of that patch would generate a query containing
"tag=''" when the alpha identifier is empty, avoiding the NPE because
the tag is set to "". The intention of that patch was to create a query
without "tag=''" if the alpha identifier is empty, and if you do that,
you will get a NPE in AdnRecord.isEqual() without this change.

- Fix bug in IccProvider.delete() where efType was compared against local
FDN constant rather than IccConstants.EF_FDN. This would always return
false. Comparing with IccConstants.EF_FDN gives the intended behavior.

Change-Id: I0ea75d7e107c7318c9a48ae6e0a15845a718f4c0