Update patch set 4
Patch Set 4: Fails; No score

(This comment applies to all your proposed changes so far, not just this specific one)

Mike, I'm afraid I'm going to be the bearer of bad news.

I've been told that your changes look like they're a solid base in the right direction.

The problem is that they touch an area of the code that is seeing some fast evolution in Google's internal source tree in order to keep up with the fast evolution of mobile graphics hardware. Our internal master tree (which is the one into which we can merge external contributions) has already drifted significantly from the state of the open-source master tree. Even if we found a way to resolve the many merge conflicts, we expect that that area of the code will undergo many more changes before it gets open-sourced, and during that time it won't be practical on our side to make sure that the functionality you're proposing will still work (in fact we're pretty sure it won't).

The best approach I could recommend would be to create a copy of our frameworks/base project into a git instance of your own (there are several good git hosting services out there), and to make all your changes on top of the android-2.1_r2.1s tag (which is the tag that OEMs are asked to use). This way, OEMs that want to use your functionality will be able to easily merge your code with theirs and ship eclair devices with your new functionality, independently of the longer-term concerns of the Android Open-Source Project. Enthusiasts will also be able to merge their changes into their homebrew systems.

Sorry I didn't get a chance to tell you earlier, I could have saved you some hard work.

Patch-set: 4
Reviewer: Gerrit User 1000660 <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=-1