Update patch set 1
Patch Set 1: No score

I do not agree with Dianne and Joe that this should be done in the HAL.  If the intent is to change the *default* values, it must be done in the framework.  The HAL can only be used to override the color in all cases and then we would break the Notification Manager APIs for controlling this (like we did with the trackball light on Nexus One).

However, system properties are not the way to do this.  The best way would be to move the current hard coded values to frameworks/base/core/res/res/values/config.xml so they can be overridden by a product specific overlay.

Patch-set: 1
Label: Verified=0