Update patch set 1
Patch Set 1: Verified; Looks good to me, approved

The description of this checkin is wrong. This does not fix the animation when the visibility changes, but when the view is removed/added from/to the window. This is very different. However, this fixes a potential bug.

Patch-set: 1
Reviewer: Gerrit User 1001146 <1001146@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2
Label: Verified=+1