Adds the ability to customize the fast scroller drawables
In order to be able to customize the fast scroller, this patch makes it
possible to set the drawables used for the thumb and the background of
the index letter via the theme. It also adds attributes for the
AbsListView which allows the drawables to be set directly in a layout
XML. Setters and getter are added as well.

Change-Id: Ia855b6b6bd91ff300a90406f0e38c04e0a2832f7