Update patch set 2
Patch Set 2: No score; Do not submit

I would certainly like to see this limitation removed, but the solution is to remove the up-front deflation of the data -- it should instead be deflated as it is streamed to the app.  I would definitely take a good change that does that.

Patch-set: 2
Reviewer: Gerrit User 1000559 <1000559@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2
Label: Verified=0