Update patch set 1
Patch Set 1: No score; Do not submit

Hi,

Thanks for your submission. There are a couple of important issues with this patch. First of all the lack of documentation, or poorly documented APIs. Your patch also does not update current.xml which will lead to a build breakage. You also did not include any test for the new feature.

Finally, and more importantly, this is not something we wish to support in the framework. You are recreating a simplistic ListView, without all of its benefits (including the recycler.)

Patch-set: 1
Reviewer: Gerrit User 1001146 <1001146@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2
Label: Verified=0