Update patch set 1
Patch Set 1:

Hi Romain,

We believe that this is a good improvement to the framework because the Gallery has functionality that ListView does not have (snap selected item to center, etc.).  There are many cases where developers will want that functionality in a vertical scrolling view.

Regarding the lack of documentation and lack of a test case.  We will provide those missing elements.  We will also update current.xml appropriately.

Thanks,
-Michael

Patch-set: 1