Corrected buffer overflow when parsing /proc/wakelocks
The android_os_Process_parseProcLineArray in android_util_Process.cpp
needs one extra byte to terminate strings. This was already handled correctly
in android_os_Process_readProcFile, but was not handled in java when
parsing /proc/wakelocks in readKernelWakelockStats in BatteryStatsImpl.java.
This caused an assert in NewStringUTF when the output from /proc/wakelocks
was larger than 4096 bytes. The buffer was also increased in order to be
able to parse all wakelocks completely.

Change-Id: I22d2feeac0b75005819ea4fd56ea7d10f47e0fc4