Update patch set 1
Patch Set 1:

Vasu, I'm not following you.  The goal of this change is primarily to allow usage of CursorJoiner on non-string columns.  The issue there is that strings and integers don't sort the same way (for instance, "2" > "12", but 2 < 12).  I don't understand what you mean by issuing a second query.

To respond to your individual points:

1) What do you mean "is not good enough"?  Is your concern that the caller may provide a buggy comparator and then try to blame the CursorJoiner?  Today they could provide unsorted cursors, breaking the joiner.  I see this as a documentation concern, not an implementation one.

2) The CursorJoiner already stipulates that the tuple of columns compared must be unique.  I have not changed this requirement.

3) Handling null columns would be up to the comparator to do if necessary.  If the schema permits nulls, then the caller must be willing to compare them as SQLite would order them.

Patch-set: 1