Update patch set 1
Patch Set 1:

> How would the CursorJoiner know which SQLite method 
> to use to read each join columns values (that is,
> when to call cursor.getString versus cursor.getLong)?
caller passes in the ObjectTypes for the columns to be compared.
best illustrated in code. will send you 2 java files - would be great to attach them right here. anyone knows how to do that?)

you will see in the attached files that most applications will be use this modified class without having to specify a comparator to handle Integer, Float etc.
 
>Also, there are suspicious performance issues with this
> approach.  First, you suffer autoboxing overhead by using
> Object[] when integer types are being used.
I wouldn't worry about the overhead of autoboxing.
much bigger overhead if the loading of a new class for the comparator (if the app uses the new constructor you proposed) 

>Second, as the CursorJoiner works today the columns are 
>first read and copied into an array, then compared.  
>This means that for each joined column there is memory
> overhead and 2 additional writes.  There is also up to
> n-1 unnecessary read/write pairs and 1 allocation where n
> is the number of join columns.

you pointed this out in the code. good one. but this issue is orthogonal to the change we are talking about, in this CL.
maybe you could submit another CL addressing this issue?

maybe you could

Patch-set: 1