Update patch set 2
Patch Set 2: No score

Well, it is hard to say in the general case, as we dont have experience with alot of different RIL implementations. The one we are currently using is dispatching CBS as raw causing the messages to become garbage in the Java layer, this is what triggered this change.

The documentation in ril.h on how these types should be interpreted in Java is not really obvious though as that documentation only refferrs to C types.

Patch-set: 2
Reviewer: Gerrit User 1002951 <1002951@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0