Added an addAll to the ArrayAdapter
It behaves the same as ArrayList.addAll, since
the Object containing the Objects in the Adapter
is a List. Now you can add multiple Objects
at once, instead of looping over a Collection
and add all items one by one.
If this change gets approved, i will add UnitTests
for the new function in the cts Project (ArrayAdapterTest.java)

Change-Id: I16f3286a8ca4cda7eb9f1a99c2aab1cc45b2550f