Update patch set 1
Patch Set 1:

I can see why it would be helpful to avoid creating custom comparators for common usage, but I don't agree that the correct solution would be to only add column type information.  For instance, it still does not address cursors which are sorted in a descending fashion.

Regarding performance concerns with the current implementation (unnecessary copies and comparisons), I agree this is a separate issue.  It should ideally be fixed first in a separate commit, but of course the changes to the API in this changeset will be affected.

Because of the custom sort order case, we still need a way to provide custom comparators.  Perhaps there is a compromise that keeps the features you mention (building the column indices, etc) outside the comparators and avoids unnecessary object creation.  With a new interface, we could define a compare function which looks like int compare(Cursor a, int[] columnsA, Cursor b, int[] columnsB).  So a sample string-based implementation could be:

private static final CursorJoinerComparator mLegacyComparator = new CursorJoinerComparator() {
    public int compare(Cursor a, int[] columnsA, Cursor b, int[] columnsB) {
        int n = columnsA.length;
        for (int i = 0; i < n; i++) {
            String aValue = a.getString(i);
            String bValue = b.getString(i);
            if (aValue == null) {
                if (bValue == null) continue;
                return -1;
            }

            if (bValue == null) {
                return 1;
            }

            int comp = aValue.compareTo(bValue);
            if (comp != 0) {
                return comp < 0 ? -1 : 1;
            }
        }
    }
};

The compare logic above is the same as compareStrings today.

Thoughts?

Patch-set: 1