Update patch set 1
Patch Set 1:

I guess you are suggested that we can treat NV_READY as ICC_READY. The only concern I had with that is that there could be other objects that register for ICC_READY which might try to do something when they receive the ICC_READY event.

Also ICC_UNUSED will be more appropriate if we consider a case where one is using NV for subscription when there is a RUIM card or SIM card in the device.

Patch-set: 1
Reviewer: Gerrit User 1003413 <1003413@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0